New arquiteture for project
